Make graph perf perf 💯
This replaces the (GPU rendered, reflow causing) height change with a GPU-accelerated, non-reflow-causing 2D transform.

Bars are not given an explicit height, but the combination of their relative scale, the maximum scale, and whether the full graph is desired. On the repaint of the "graph," they will enter their full scales, transforming from the initial downscaled sizes (calculated to be the graph's minimum height) to full size.

This appears, in dev, to reduce jank, and make the page smoother on load. Hopefully it will for you, too.

In doing this, it breaks out some of the (now-)shared pipeline constants into their own constant file.