Lossless emoji replacement
This updates the Emoji class to produce copyable output which, in all cases, would produce the same output if put through the parser again.

The alt text for non-unicode emoji is now shown as the same colon-wrapped version we accept as input, whilst Unicode emoji continues to use the Unicode representation.