Make navigation scrollbars lazy
When a use has scrollbars forced on (mouse is plugged in, using a tablet) and `overflow` properties are set to `scroll`, scrollbars are always shown. On OSX by default, most scrollbars sit above the content, whereas when they are shown they collapse the content. With this change, scrollbars are shown when the screen is small and the content is wide, but not triggered in the default case.