Fix AgentsCount inconsistencies
This updates the AgentsCount component to be Relay-aware, and instead of alternatingly trusting the Pusher agent count and Relay agent count, instead takes a change on the Pusher end as a cue to `forceFetch` updated values from Relay.

This causes the component to stay pretty stable when it comes to repaints.

I also took the opportunity to switch over to using `shallowCompare` in `shouldComponentUpdate` and remove the component from the Webpack bridge in app.js

Fixes https://3.basecamp.com/3453178/buckets/1445180/todos/323608085