Make the Changelogs badge bulletproof
Previously, it suffered from the same state caching issues as other Pusher-powered things that could also be interacted with (opening, for instance, the Organizations dropdown, would cause it to use the cached state), this updates it to act correctly in all cases.

It also makes it a relay container which accepts the isMounted variable from its parent, so that it can force an update of only its data set whilst also being updated on-mount without triggering a separate request for its data!