Improve duration formatting
A bug existed which meant the most significant unit of duration would be shown including the value of next most significant unit of duration. This caused strangely divergent results between `micro` and `full` date rendering.

This improves the logic which controls truncation and merging of duration segments, and fixes #165's issue.