Rename the Autocomplete field
The Autocomplete Field does not act like other components prefixed with `Form-` (which as far as I can tell are intended as a consistent set of things for labeled, feedback-providing form controls), and that's currently the intended behaviour.

This, therefore, renames it to `AutocompleteField` to remove the implication that they're related.