Improve Dialog behaviour
This allows the dialog to scroll. What it does is rearrange the dialog to allow the entire dialog to appear partially offscreen, and the rest to scroll into view.

This has a couple of caveats, which might be acceptable or might need working around.

In any case, this is intended in support of #222, which can in some cases create dialogs which appear partially offscreen, and currently can't be scrolled to.

The caveats;
* Scrolling the dialog requires the cursor to be over the dialog, not the dialog backdrop
  * This is because the scroll container is transparent so the scroll event is passed to the backdrop, which is not part of a scroll container, and thus passed to the page.
  * We can work around this by putting the backdrop in the scroll container, but we need to think more about how we manage the transition animations if we do that.
* Firefox collapses the padding at the bottom of the flex container.

The easiest way to check out this behaviour is to open the Support dialog and shrink the window to ~320px wide and ~very smol high!