Powered by Relay Modern (but not really)
This bumps us to the 1.0.0 release of relay in classic mode. This works
exactly the same as before, but we can start to use relay modern style
containers using relay compat:

https://facebook.github.io/relay/docs/relay-compat.html

Eventually when everything's switched to compat and we've switched to a
QueryRenderer instead of react-router-relay we'll be able to switch
properly to modern. 🎉