Use browserHistory events for flash clearing instead of the router
I discovered that `onUpdate` actually runs when the page loads, as well
as on every route change. This means that we end up clearing flashes
when page loads, and in some cases - they don't even show at all
(depending on how fast the browser is).

By using `browserHistory` we _only_ get events when the route is changed
in Javascript, not when the page loads :ok_hand: