Hide HTTP Basic Auth passwords in Pipeline Header
This change censors passwords included in repository URLs when they're shown in the Pipeline header.

I noticed that a few customers use HTTP Basic Auth for git authentication, and while that's not dangerous in and of itself, the fact that it means their credentials are shown in the top of their (and my, if I'm doing support for them!) screen on every page made me nervous.

To that end, this change makes any passwords included in a URL be replaced with up to 8 • characters. This doesn't stop people using HTTP basic auth, and hopefully the dots are well-understood enough as meaning "hidden password."

The password can of course still be harvested from the page, either via the JavaScript or GraphQL, but discouraging encoding passwords in our web UI is probably a separate change. :)