Allow for graceful Agent stopping
This changes the behaviour of the existing Agents page, which treats the "Stop" button as an immediate, ungraceful, force stop. This means we don't have any user-facing way to gracefully stop an agent!

Following this change, the "Stop Agent" button is a graceful stop. We'll ask the agent politely to stop.

If the Agent takes more than ten (±3) seconds to stop, we then show a "Force Stop Agent" button in its place. This warns that it'll cancel the running job (as the current warning does) and will function the same as our current behaviour.