Make sure HMR always works
Webpack dev server seems to serve slightly different assets to what are
compiled by messing with them in middleware, and it needs to talk
directly to its own host because it uses socket io or something to do
websocket-y hot module reloading. So make sure it always spits out a
manifest with a fully-qualified public path.