Remove webpack config sha prefix
This needs to be removed to do variable frontend paths. It's part of the
public path which is being replaced on the fly, and we don't store this
prefix anywere else useful.

I remember we did this because webpack builds weren't stable after
config changes. I think maybe something to do with the way the md5
extension thing was doing hashing of the assets. But the [chunkhash]
is based on the content of the asset and avoid that whole problem, so
I'm pretty sure we don't need it any more.